Docker Cmpose + major functional & style additions The web app has been configured to run as a multi-container application using docker compose. A docker-compose.yml configuration file has been created to define the containers as well as set up any port mapping, directory mounting, and persistent volumes required, as well as environment variables required for the MySQL database.
Directories have been added and files have been
arranged into the directory associated with their
container and its role in the web application.
The database folder contains the database initialisation file,
and a testdata.sql file for populating the database with some starting data.
The other files for the node.js web application being placed in
a directory named event_planner, which also includes a Dockerfile
used to install the necessary dependencies using npm,
as well as start the express server within the application's container.

Large functionality & styling fixes and additions have been made as well.

APP CHANGELOG:

- The AJAX function for login was getting interrupted by a redirect,
reloading the page before it could finish. This has been fixed by removing the
redirect route from the form action, and by preventing the default
form submit behaviour which reloads the page.
- The signup form was modified to also trigger its associated AJAX
function on form submit and preventing default submit behaviour,
rather than a button onclick event, for the sake of consistency.
- Google OAuth origin mismatch error was fixed so that localhost
and the ports 80, 3000, and 8080 are a valid origin for Google OAuth.
- Styling of the login and signup forms was improved, with better
margins, paddings, centering, and button colour and radius.
- Form labels are now above the input fields instead of below.

- Updates were made to a few important outdated node packages:
-- The Mysql package was changed to the MySql2 package as it provides
newer authentication methods required for the newer MySQL Version (9)
being used in the database container.
-- The google-auth-library package was updated to the newest version (10.4.0),
and the application was changed from using Google Sign-in to instead using
Google Identity Services, due to Google Sign-in no longer being allowed for OAuth
client ids created after July 2022.

- Made many changes to css stylesheets to fix incorrect references to elements,
such as using id instead of classname, and rearranged some ids and classnames
to decrease instances of redundant styling.
- Changed styling for the navbar so that nav elements related to login and signup use
a blue, styled button similar to the ones used on the forms, while other nav elements
are made to appear as clickable text rather than buttons.
- Fixed styling of the searchbar, and added a blue button on the end of the search input
containing a search icon inside. Also changed the length so that it displays better on mobile devices.
- Aligned the dropdown input to the right side of the screen instead of on the end of the searchbar.

- Improved the styling for various event-related elements.
-- The event creation form now has icons relevant to the input fields, instead of just user icons.
-- The icons have been positioned better alongside the input field.
-- Form labels are now above the input fields for readability and consistency with other forms.
-- Fixed the styling for buttons for availability and event time inputs.
-- Changed the styling for the whitespace property to pre-wrap on pre elements
for better mobile device formatting, while still preserving whitespace.

- Fixed error in database creation SQL for availability table marking the userID itself as unique,
preventing user from specifying availability for other events.
- Fixed the route for adding availabilities to database by having the table be updated instead of an insertion being made
when an availability for that event and user already exists within the table.
- Event creation form also now properly allows time to be set upon event creation,
instead of requiring it to be set afterwards in the myevents page.

- Event host is now also displayed on the event by modifying the sql query to include an innerjoin with the user table.
- Event time inputs now display the times within them if they have been previously set.
This included making a function that is called when getting events which uses the fetch api along with
the await operator to wait for return of the user's availability data.
- The previous function for getting event availabilities, getAvailability(), was renamed to getEventAvailability, to
avoid confusion with the new function for getting user availabilites, getUserAvailability.

- Fixed the displaying of availabilities for events in the myevents page.
-- Availabilities are now displayed at the bottom of the associated event
instead of at the very bottom of the page.
--Clicking the button several times no longer creates additional copies of the availabilities.
-- Fixed the hideDiv function so that the availabilities div is now destroyed instead of
being made invisible, meaning that there aren't needless lingering copies each time that can't be accessed.
-- Added message in the bottom right corner to make it clearer to the user how to hide the availabilities.

- Fixed dynamically-created, event-related html elements being given the same id,
(which was also being used to select the element). This was done by appending the eventID onto the id string.

- Commented out some redundant javascript functions (loginGoogle and SendID), instead incorporating the sendID function
into the function it was used for, getEventAvailability, by using url args for the GET request in getEventAvailability.

- Time and Date formats changed.
-- Times are now displayed as HH:MM instead of HH:MM:SS.
-- Dates are now displayed as DD/MM/YYYY instead of YYYY-MM-DD
-- The "Next Upcoming Event" on the myevents page has been
reformatted and now includes the day of the week, and the word "at"
as a separator between the date and time for easier readability.
-- The date is also now included on all the events listed in the myevents page,
with only one date being shown on events that start and end on the same day.

- Events are now no longer displayed after their end date has elapsed.

- All uses of the innerHTML property for dynamically setting the text of an element have
been changed to use innerText instead to prevent possible Stored Cross-Site Scripting (XSS) exploitation.

- A favicon has been added for the site. The favicon is a smaller version of the calendar.png image used on the site's front page.